+++ shell/components/Drawer/ResourceDetailDrawer/ConfigTab.vue
@@ -1,17 +1,11 @@
-<script lang="ts">
+<script setup lang="ts">
 import { useI18n } from '@shell/composables/useI18n';
 import { _VIEW } from '@shell/config/query-params';
 import { useStore } from 'vuex';
 import Tab from '@shell/components/Tabbed/Tab.vue';
+import { ConfigProps } from '@shell/components/Drawer/ResourceDetailDrawer/types';
 
-export interface Props {
-  resource: any;
-  component: any;
-  resourceType: string;
-}
-</script>
-<script setup lang="ts">
-const props = defineProps<Props>();
+const props = defineProps<ConfigProps>();
 const store = useStore();
 const i18n = useI18n(store);
 </script>
