+++ shell/components/Drawer/types.ts
@@ -1,0 +1,3 @@
+export interface Props {
+    ariaTarget: string;
+}
